Split Footer into two files
