+++ frontend/src/App.js
@@ -15,7 +15,8 @@ import config from './config'
 import {routeTo} from './helpers/routes'
 import {provideIntl} from './components/HOC/intl'
 import {provideTheme, withTheme, THEME_DEFINITIONS} from './components/HOC/theme'
-import {Navbar, MobileNavbar, Footer, Link} from './components/visual'
+import {Navbar, MobileNavbar, Link} from './components/visual'
+import Footer from './screens/Footer'
 import {useI18n, InjectHookIntlContext} from '@/i18n/helpers'
 import {AutoSyncProvider} from './screens/Staking/context/autoSync'
