+++ frontend/src/components/visual/index.js
@@ -6,7 +6,6 @@ export {default as ErrorOverlay} from './ErrorOverlay'
 export {default as DebugApolloError} from './DebugApolloError'
 export {default as SimpleLayout} from './SimpleLayout'
 export {default as SummaryCard} from './SummaryCard'
-export {default as Footer} from './Footer'
 export {default as Button} from './Button'
 export {Navbar, MobileNavbar} from './Navbar'
 export {default as EntityIdCard, EntityCardContent, EntityCardShell} from './EntityIdCard'
