+++ frontend/src/i18n/locales/en.json
@@ -38,17 +38,6 @@
   "components.visual.Alert.noResults": "No results",
   "components.visual.Alert.warning": "Warning",
   "components.visual.ComparisonMatrix.copyText": "Copy",
-  "components.visual.Footer.copyright": "All rights reserved",
-  "components.visual.Footer.emailButton": "Email",
-  "components.visual.Footer.genericError": "Sorry. An error occurred. Please try again later.",
-  "components.visual.Footer.invalidEmail": "Incorrect format of email",
-  "components.visual.Footer.privacyLink": "Privacy and Cookie policy",
-  "components.visual.Footer.subscribeButton": "Subscribe",
-  "components.visual.Footer.subscribeHeader": "Let's stay in touch!",
-  "components.visual.Footer.subscribeInfo": "By subscribing, you agree to receive news and updates about our new features and products in accordance with our {privacyLink}.",
-  "components.visual.Footer.subscribeSuccess": "You are successfully subscribed!",
-  "components.visual.Footer.subscribeText": "Subscribe to Seiza to receive news and updates about staking, rewards and new features!",
-  "components.visual.Footer.subscribeToNewsletter": "Subscribe to newsletter",
   "components.visual.LoadingError.GENERIC": "There was an error loading the data",
   "components.visual.LoadingError.NETWORK_ERROR": "We have trouble communicating with the backend",
   "components.visual.LoadingError.NOT_FOUND__ADDRESS": "Address does not exist",
@@ -238,6 +227,21 @@
   "screens.Blockchain.Transaction.index.size": "Size:",
   "screens.Blockchain.Transaction.index.slot": "Slot:",
   "screens.Blockchain.Transaction.index.transactionId": "Transaction Id",
+  "screens.Footer.MainFooter.copyright": "All rights reserved",
+  "screens.Footer.MainFooter.privacyLink": "Privacy and Cookie policy",
+  "screens.Footer.MainFooter.subscribeInfo": "By subscribing, you agree to receive news and updates about our new features and products in accordance with our {privacyLink}.",
+  "screens.Footer.MainFooter.subscribeToNewsletter": "Subscribe to newsletter",
+  "screens.Footer.SubscribeFooter.copyright": "All rights reserved",
+  "screens.Footer.SubscribeFooter.emailButton": "Email",
+  "screens.Footer.SubscribeFooter.genericError": "Sorry. An error occurred. Please try again later.",
+  "screens.Footer.SubscribeFooter.invalidEmail": "Incorrect format of email",
+  "screens.Footer.SubscribeFooter.privacyLink": "Privacy and Cookie policy",
+  "screens.Footer.SubscribeFooter.subscribeButton": "Subscribe",
+  "screens.Footer.SubscribeFooter.subscribeHeader": "Let's stay in touch!",
+  "screens.Footer.SubscribeFooter.subscribeInfo": "By subscribing, you agree to receive news and updates about our new features and products in accordance with our {privacyLink}.",
+  "screens.Footer.SubscribeFooter.subscribeSuccess": "You are successfully subscribed!",
+  "screens.Footer.SubscribeFooter.subscribeText": "Subscribe to Seiza to receive news and updates about staking, rewards and new features!",
+  "screens.Footer.SubscribeFooter.subscribeToNewsletter": "Subscribe to newsletter",
   "screens.Home.Charts.index.adaInBillions": "{count}B",
   "screens.Home.Charts.index.adaSent": "Total Ada Sent",
   "screens.Home.Charts.index.day": "Day (UTC)",
