+++ frontend/src/i18n/locales/ja-JP.json
@@ -38,17 +38,6 @@
   "components.visual.Alert.noResults": "結果がありません",
   "components.visual.Alert.warning": "注意",
   "components.visual.ComparisonMatrix.copyText": "コピー",
-  "components.visual.Footer.copyright": "転載禁止",
-  "components.visual.Footer.emailButton": "メールアドレス",
-  "components.visual.Footer.genericError": "エラーが発生しました。もう一度お試しください",
-  "components.visual.Footer.invalidEmail": "メールアドレスの形式が正しくありません",
-  "components.visual.Footer.privacyLink": "プライバシー＆クッキーポリシー",
-  "components.visual.Footer.subscribeButton": "登録",
-  "components.visual.Footer.subscribeHeader": "メールマガジンの登録",
-  "components.visual.Footer.subscribeInfo": "購読することで、当社の{privacyLink}に従って新しいフィーチャーや製品に関するニュースや最新情報を受け取ることに同意します。",
-  "components.visual.Footer.subscribeSuccess": "登録が完了しました",
-  "components.visual.Footer.subscribeText": "Seizaに登録して、ステーキングや報酬、新機能に関する最新情報を受け取りましょう。",
-  "components.visual.Footer.subscribeToNewsletter": "メールマガジンの登録をする",
   "components.visual.LoadingError.GENERIC": "データの読み込み時にエラーが発生しました",
   "components.visual.LoadingError.NETWORK_ERROR": "バックエンドとの通信中に問題が発生しました",
   "components.visual.LoadingError.NOT_FOUND__ADDRESS": "アドレスが存在しません",
@@ -238,6 +227,21 @@
   "screens.Blockchain.Transaction.index.size": "サイズ:",
   "screens.Blockchain.Transaction.index.slot": "スロット:",
   "screens.Blockchain.Transaction.index.transactionId": "取引ID",
+  "screens.Footer.MainFooter.copyright": "転載禁止",
+  "screens.Footer.MainFooter.privacyLink": "プライバシー＆クッキーポリシー",
+  "screens.Footer.MainFooter.subscribeInfo": "購読することで、当社の{privacyLink}に従って新しいフィーチャーや製品に関するニュースや最新情報を受け取ることに同意します。",
+  "screens.Footer.MainFooter.subscribeToNewsletter": "メールマガジンの登録をする",
+  "screens.Footer.SubscribeFooter.copyright": "転載禁止",
+  "screens.Footer.SubscribeFooter.emailButton": "メールアドレス",
+  "screens.Footer.SubscribeFooter.genericError": "エラーが発生しました。もう一度お試しください",
+  "screens.Footer.SubscribeFooter.invalidEmail": "メールアドレスの形式が正しくありません",
+  "screens.Footer.SubscribeFooter.privacyLink": "プライバシー＆クッキーポリシー",
+  "screens.Footer.SubscribeFooter.subscribeButton": "登録",
+  "screens.Footer.SubscribeFooter.subscribeHeader": "メールマガジンの登録",
+  "screens.Footer.SubscribeFooter.subscribeInfo": "購読することで、当社の{privacyLink}に従って新しいフィーチャーや製品に関するニュースや最新情報を受け取ることに同意します。",
+  "screens.Footer.SubscribeFooter.subscribeSuccess": "登録が完了しました",
+  "screens.Footer.SubscribeFooter.subscribeText": "Seizaに登録して、ステーキングや報酬、新機能に関する最新情報を受け取りましょう。",
+  "screens.Footer.SubscribeFooter.subscribeToNewsletter": "メールマガジンの登録をする",
   "screens.Home.Charts.index.adaInBillions": "{count}B",
   "screens.Home.Charts.index.adaSent": "送信済みADA合計",
   "screens.Home.Charts.index.day": "日付(UTC)",
